move to right directory
